+++ miniapp/src/test/java/com/rakuten/tech/mobile/miniapp/js/hostenvironment/HostEnvironmentInfoBridgeSpec.kt
@@ -1,103 +1,94 @@
 package com.rakuten.tech.mobile.miniapp.js.hostenvironment
 
+import android.os.Build
+import androidx.test.core.app.ActivityScenario
+import androidx.test.ext.junit.runners.AndroidJUnit4
 import com.google.gson.Gson
 import com.rakuten.tech.mobile.miniapp.*
 import com.rakuten.tech.mobile.miniapp.TEST_CALLBACK_ID
-import com.rakuten.tech.mobile.miniapp.TEST_MA
-import com.rakuten.tech.mobile.miniapp.display.WebViewListener
-import com.rakuten.tech.mobile.miniapp.errors.MiniAppBridgeErrorModel
 import com.rakuten.tech.mobile.miniapp.js.*
 import org.amshove.kluent.When
 import org.amshove.kluent.calling
 import org.amshove.kluent.itReturns
 import org.junit.Before
 import org.junit.Test
+import org.junit.runner.RunWith
 import org.mockito.Mockito
 import org.mockito.kotlin.mock
+import org.mockito.kotlin.times
 import org.mockito.kotlin.verify
 
-class HostEnvironmentInfoBridgeSpec {
-    private lateinit var miniAppBridge: MiniAppMessageBridge
-    private val hostEnvCallbackObj = CallbackObj(
-            action = ActionType.GET_HOST_ENVIRONMENT_INFO.action,
-            param = null,
-            id = TEST_CALLBACK_ID
-    )
-
-    private val webViewListener: WebViewListener = mock()
-    private val bridgeExecutor = Mockito.spy(MiniAppBridgeExecutor(webViewListener))
-
-    private val hostEnvironmentInfo = HostEnvironmentInfo("Android 30", "012345", "X.X.X")
-    private val infoErrMessage = "{\"type\":\"${HostEnvironmentInfoBridge.ERR_GET_ENVIRONMENT_INFO} $TEST_ERROR_MSG\"}"
-    private val testEnvironmentError = HostEnvironmentInfoError(null, infoErrMessage)
-    private fun createEnvironmentImpl(canGetInfo: Boolean, isImplemented: Boolean): HostEnvironmentBridgeDispatcher {
-        return if (isImplemented) {
-            object : HostEnvironmentBridgeDispatcher {
-                override fun getHostEnvironmentInfo(
-                    onSuccess: (info: HostEnvironmentInfo) -> Unit,
-                    onError: (infoError: HostEnvironmentInfoError) -> Unit
-                ) {
-                    if (canGetInfo)
-                        onSuccess.invoke(hostEnvironmentInfo)
-                    else
-                        onError.invoke(testEnvironmentError)
-                }
-            }
-        } else {
-            object : HostEnvironmentBridgeDispatcher {}
-        }
-    }
+@RunWith(AndroidJUnit4::class)
+class HostEnvironmentInfoBridgeSpec : BridgeCommon() {
+    private val miniAppBridge = Mockito.spy(createDefaultMiniAppMessageBridge())
 
     @Before
     fun setup() {
-        miniAppBridge = Mockito.spy(object : MiniAppMessageBridge() {})
         When calling miniAppBridge.createBridgeExecutor(webViewListener) itReturns bridgeExecutor
         miniAppBridge.init(
                 activity = TestActivity(),
                 webViewListener = webViewListener,
                 customPermissionCache = mock(),
                 downloadedManifestCache = mock(),
-                miniAppId = TEST_MA.id,
+                miniAppId = TEST_MA_ID,
                 ratDispatcher = mock()
         )
     }
 
-    @Test
-    @SuppressWarnings("MaximumLineLength")
-    fun `postError should be called when there is no get environment info retrieval implementation`() {
-        val dispatcher = Mockito.spy(createEnvironmentImpl(false, false))
-        miniAppBridge.setHostEnvironmentBridgeDispatcher(dispatcher)
-        val errMsg = "{\"type\":\"${HostEnvironmentInfoBridge.ERR_GET_ENVIRONMENT_INFO}" +
-                " ${ErrorBridgeMessage.NO_IMPL}\"}"
-        miniAppBridge.postMessage(Gson().toJson(hostEnvCallbackObj))
-
-        verify(bridgeExecutor).postError(hostEnvCallbackObj.id, errMsg)
-    }
-
-    @Test
-    fun `postError should be called when hostapp doesn't providing environment info`() {
-        val dispatcher = Mockito.spy(createEnvironmentImpl(false, true))
-        val userInfoBridgeWrapper = Mockito.spy(createBridgeWrapper(dispatcher))
-        userInfoBridgeWrapper.onGetHostEnvironmentInfo(hostEnvCallbackObj.id)
-
-        verify(bridgeExecutor).postError(hostEnvCallbackObj.id, Gson().toJson(
-                MiniAppBridgeErrorModel(message = infoErrMessage)
-        ))
-    }
+    private val hostEnvCallbackObj = CallbackObj(
+            action = ActionType.GET_HOST_ENVIRONMENT_INFO.action,
+            param = null,
+            id = TEST_CALLBACK_ID
+    )
+    private val hostEnvJsonStr = Gson().toJson(hostEnvCallbackObj)
 
     @Test
     fun `postValue should be called when retrieve environment info successfully`() {
-        val dispatcher = Mockito.spy(createEnvironmentImpl(true, true))
-        val userInfoBridgeWrapper = Mockito.spy(createBridgeWrapper(dispatcher))
-        userInfoBridgeWrapper.onGetHostEnvironmentInfo(hostEnvCallbackObj.id)
-
-        verify(bridgeExecutor).postValue(hostEnvCallbackObj.id, Gson().toJson(hostEnvironmentInfo))
+        ActivityScenario.launch(TestActivity::class.java).onActivity { activity ->
+            val hostEnvironmentInfo = HostEnvironmentInfo(
+                    platformVersion = Build.VERSION.RELEASE,
+                    hostVersion = activity.packageManager.getPackageInfo(activity.packageName, 0).versionName,
+                    sdkVersion = BuildConfig.VERSION_NAME
+            )
+            val miniAppBridge = Mockito.spy(createDefaultMiniAppMessageBridge())
+            When calling miniAppBridge.createBridgeExecutor(webViewListener) itReturns bridgeExecutor
+            miniAppBridge.init(
+                    activity = activity,
+                    webViewListener = webViewListener,
+                    customPermissionCache = mock(),
+                    downloadedManifestCache = mock(),
+                    miniAppId = TEST_MA_ID,
+                    ratDispatcher = mock()
+            )
+            miniAppBridge.postMessage(hostEnvJsonStr)
+            verify(bridgeExecutor).postValue(hostEnvCallbackObj.id, Gson().toJson(hostEnvironmentInfo))
+        }
     }
 
-    private fun createBridgeWrapper(dispatcher: HostEnvironmentBridgeDispatcher): HostEnvironmentInfoBridge {
-        val wrapper = HostEnvironmentInfoBridge()
-        wrapper.setMiniAppComponent(bridgeExecutor)
-        wrapper.setHostEnvironmentBridgeDispatcher(dispatcher)
-        return wrapper
+    @Test
+    fun `postError should be called when cannot retrieve environment info`() {
+        val miniAppBridge = Mockito.spy(createMiniAppMessageBridge(false))
+        val infoErrMessage = "{\"type\":\"{\\\"type\\\":\\\"Cannot get host environment info: error_message\\\"}\"}"
+        val webViewListener = createErrorWebViewListener(infoErrMessage)
+        val bridgeExecutor = Mockito.spy(miniAppBridge.createBridgeExecutor(webViewListener))
+        When calling miniAppBridge.createBridgeExecutor(webViewListener) itReturns bridgeExecutor
+        miniAppBridge.init(
+                activity = TestActivity(),
+                webViewListener = webViewListener,
+                customPermissionCache = mock(),
+                downloadedManifestCache = mock(),
+                miniAppId = TEST_MA_ID,
+                ratDispatcher = mock()
+        )
+        miniAppBridge.postMessage(hostEnvJsonStr)
+
+        verify(bridgeExecutor).postError(TEST_CALLBACK_ID, infoErrMessage)
+    }
+
+    @Test
+    fun `postValue should not be called when using default method without Activity`() {
+        miniAppBridge.postMessage(hostEnvJsonStr)
+
+        verify(bridgeExecutor, times(0)).postValue(TEST_CALLBACK_ID, "anyValue")
     }
 }